Back-End/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy>=2.1.0
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1
requests==2.31.0
